finished the acces to Product

+++ src/main/java/com/example/Ecomerce/feature1/Service/ProduitServiceImp.java
@@ -1,30 +1,40 @@
 package com.example.Ecomerce.feature1.Service;
 
 import com.example.Ecomerce.feature1.Model.Produit;
+import com.example.Ecomerce.feature1.Model.Utilsateur;
 import com.example.Ecomerce.feature1.Repository.ProduitRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.rest.webmvc.ResourceNotFoundException;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
 @Service
-
+@Transactional
 public class ProduitServiceImp implements IProduitService {
     @Autowired
     private ProduitRepository produitRepository;
+    @Autowired ProduitSecurityService produitSecurityService;
     @Override
-    public Produit Saveproduit(Produit produit) {
+    public Produit Saveproduit(Produit produit, Authentication authentication) {
         return produitRepository.save(produit);
     }
 
     @Override
-    public void Deleteproduit(Long id) {
+    public void Deleteproduit(Long id,Authentication authentication) {
+        if (!produitSecurityService.isOwnerOrAdmin(id, authentication)) {
+            throw new RuntimeException("Vous n'avez pas les droits pour supprimer ce produit");
+        }
         produitRepository.deleteById(id);
     }
 
     @Override
-    public Produit Updateproduit(Long id, Produit produit) {
+    public Produit Updateproduit(Long id, Produit produit,Authentication authentication) {
+        if (!produitSecurityService.isOwnerOrAdmin(id, authentication)) {
+            throw new RuntimeException("Vous n'avez pas les droits pour modifier ce produit");
+        }
         Optional<Produit> produitOptional = produitRepository.findById(id);
 
         if (produitOptional.isPresent()) {
